Rewrite API to support large files > INT_MAX. Add QFileBuffer to work with large files

+++ document/buffer/qhexbuffer.cpp
@@ -3,10 +3,10 @@
 
 QHexBuffer::QHexBuffer(QObject *parent) : QObject(parent) { }
 
-uchar QHexBuffer::at(int idx) { return this->read(idx, 1)[0]; }
+uchar QHexBuffer::at(qint64 idx) { return this->read(idx, 1)[0]; }
 bool QHexBuffer::isEmpty() const { return this->length() <= 0; }
 
-void QHexBuffer::replace(int offset, const QByteArray &data)
+void QHexBuffer::replace(qint64 offset, const QByteArray &data)
 {
     this->remove(offset, data.length());
     this->insert(offset, data);
